Move hashers to common.py
