+++ hwilib/common.py
@@ -1,3 +1,5 @@
+import hashlib
+
 from enum import Enum
 
 from typing import Union
@@ -40,3 +42,19 @@ class AddressType(Enum):
             return AddressType[s.upper()]
         except KeyError:
             return s
+
+
+def sha256(s: bytes) -> bytes:
+    return hashlib.new('sha256', s).digest()
+
+
+def ripemd160(s: bytes) -> bytes:
+    return hashlib.new('ripemd160', s).digest()
+
+
+def hash256(s: bytes) -> bytes:
+    return sha256(sha256(s))
+
+
+def hash160(s: bytes) -> bytes:
+    return ripemd160(sha256(s))
